+++ lib/screens/splash_screen.dart
@@ -2,6 +2,7 @@ import 'package:after_layout/after_layout.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:hackernews/config/size_config.dart';
+import 'package:hackernews/controllers/search_results_controller.dart';
 import 'package:hackernews/routes/routes.dart' as routes;
 import 'package:hackernews/widgets/shared/loading_indicator.dart';
 
@@ -13,10 +14,14 @@ class SplashScreen extends StatefulWidget {
 class _SplashScreenState extends State<SplashScreen> with AfterLayoutMixin {
   @override
   void afterFirstLayout(BuildContext context) {
+    Get.lazyPut(
+      () => SearchResultsController(),
+    );
     Future.delayed(
         Duration(
           seconds: 2,
         ), () {
+      Get.find<SearchResultsController>().init();
       Get.toNamed(
         routes.HOME,
       );