+++ lib/services/postings_service.dart
@@ -1,12 +1,12 @@
 import 'package:dio/dio.dart';
 import 'package:hackernews/config/constants.dart';
+import 'package:hackernews/models/posting_detail_model.dart';
 import 'package:hackernews/models/posting_model.dart';
-import 'package:hackernews/models/postings_model.dart';
 
 class PostingsService {
   Dio _dio = Dio();
 
-  Future<Posting> getPosting({
+  Future<PostingDetails> getPosting({
     int postingId,
   }) async {
     final String url = '$ENDPOINT/items/$postingId';
@@ -19,13 +19,13 @@ class PostingsService {
       ),
     );
     if (response.data != null) {
-      return Posting.fromJson(response.data);
+      return PostingDetails.fromJson(response.data);
     } else {
       throw response.statusMessage;
     }
   }
 
-  Future<List<Postings>> getPostingsBySearch({
+  Future<List<Posting>> getPostingsBySearch({
     String searchText,
   }) async {
     final String url = '$ENDPOINT/search';
@@ -42,7 +42,7 @@ class PostingsService {
     );
     if (response.data != null) {
       final List data = response.data['hits'] as List;
-      return data.map((e) => Postings.fromJson(e)).toList();
+      return data.map((e) => Posting.fromJson(e)).toList();
     } else {
       throw response.statusMessage;
     }